// src/enums/auth.ts
export enum LOGINS {
  DOCTOR = 'doctor',
  OPTOMETRIST = 'optometrist',
  READER = 'reader',
  PRACTICE = 'practice',
}

export enum GENDER {
  male = 'male',
  female = 'female',
}

export enum ETHNICITY {
  caucasian = 'caucasian',
  east_asian = 'east_asian',
  african = 'african',
  mixed_caucasian_east_asian = 'mixed_caucasian_east_asian',
  mixed_caucasian_african = 'mixed_caucasian_african',
  mixed_east_asian_east_asian = 'mixed_east_asian_east_asian',
  mixed_caucasian_other_non_relevant_ethnicity = 'mixed_caucasian_other_non_relevant_ethnicity',
  mixed_east_asian_other_non_relevant_ethnicity = 'mixed_east_asian_other_non_relevant_ethnicity',
  mixed_african_other_non_relevant_ethnicity = 'mixed_african_other_non_relevant_ethnicity',
  non_relevant_ethnicity = 'non_relevant_ethnicity',
}